Module tower_fault_injector::error

Error injection for tower

Layer that injects errors randomly into a service. When an error is injected, the underlying service is not called.

Usage

use tower_fault_injector::error::ErrorLayer;
use tower::{service_fn, ServiceBuilder};

// Initialize an ErrorLayer with a 10% probability of returning
// an error.
let error_layer = ErrorLayer::new(0.1, |_: &MyRequest| String::from("error"));

let service = ServiceBuilder::new()
    .layer(error_layer)
    .service(service_fn(my_service));

Decider

The decider is used to determine if a latency should be injected or not. This can be a boolean, float, Bernoulli distribution, a closure, or a custom implementation of the Decider trait.

For more information, see the decider module.

use tower_fault_injector::error::ErrorLayer;

// Never inject an error.
ErrorLayer::new(false, |_: &MyRequest| String::from("error"));
// Always inject an error.
ErrorLayer::new(true, |_: &MyRequest| String::from("error"));

// Inject an error 30% of the time.
ErrorLayer::new(0.3, |_: &MyRequest| String::from("error"));

// Inject an error based on the request content.
ErrorLayer::new(|req: &MyRequest| req.value % 2 == 0, |_: &MyRequest| String::from("error"));

Generator

The generator is a function that returns an error based on the request.

use tower_fault_injector::error::ErrorLayer;

// Customize the error based on the request payload
ErrorLayer::new(false, |req: &MyRequest| format!("value: {}", req.value));

Structs

ErrorLayer

Layer that randomly trigger errors for the service.

ErrorService

Service that randomly trigger errors instead of calling the underlying service.